Separate topology tests into separate file
