+++ ydb/core/kqp/ut/join/ya.make
@@ -25,6 +25,7 @@ SRCS(
     kqp_join_ut.cpp
     kqp_join_order_ut.cpp
     kqp_join_topology_generator.cpp
+    kqp_join_topology_ut.cpp
 )
 
 PEERDIR(
